requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
APScheduler==3.10.4
fpdf2==2.7.5